feat: 同步 OpenAPI 规范并重构 v1 代码生成器 (script4) (#3)
* feat: 同步 Amazon Ads API v1 最新规范并重新生成客户端与数据模型 (2026-08-20)

- 同步 2026-08-20 官方 OpenAPI v1 规范，合并 24 个 DSP/SB 实体至 ALL 全产品规范
- 优化 script2/generate.py 生成器，增加废弃子模块客户端文件自动清理逻辑
- 重新生成 v1 顶层通用客户端 (client/v1/) 与数据模型 (models/v1/)
- 格式化代码并通过 ruff、mypy 类型检查与 pytest 测试套件

* feat: 同步 Amazon Ads API v0 (Ads Data Manager) 最新规范并更新数据模型

* fix: 修复代码生成器将 ID 字段错误映射为 float 的问题，统一修正为 int

* feat: 使用 script4 替代 script2，基于官方 Merged OpenAPI 重构 v1 代码生成

- 下载 7 个官方 Merged OpenAPI 规范文件替代原先分散在 36 个目录的 100+ 个碎片文件
- 修复产品归属，将原本误挂在顶层的 24 个实体正确归入 SB (12 个) 与 DSP (21 个) 命名空间
- 彻底废弃并删除 script2，更新 pyproject.toml、AGENTS.md、README.md 与相关文档
- 重新生成全量 v1 Client 与 Models，通过 ruff、black、mypy 与 pytest 校验

+++ src/ads_api/client/v0/sd/budget_rules.py
@@ -29,19 +29,19 @@ class BudgetRules(BaseResource):
 
     @overload
     async def create_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["pydantic"] = "pydantic"
+        self, campaign_id: int, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["pydantic"] = "pydantic"
     ) -> CreateAssociatedBudgetRulesResponse: ...
     @overload
     async def create_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["dict"]
+        self, campaign_id: int, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["dict"]
     ) -> dict[str, Any]: ...
     @overload
     async def create_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["raw"]
+        self, campaign_id: int, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["raw"]
     ) -> httpx.Response: ...
     async def create_associated_budget_rules_for_sd_campaigns(
         self,
-        campaign_id: float,
+        campaign_id: int,
         body: CreateAssociatedBudgetRulesRequest,
         *,
         mode: Literal["pydantic", "dict", "raw"] = "pydantic",
@@ -73,18 +73,18 @@ class BudgetRules(BaseResource):
 
     @overload
     async def disassociate_associated_budget_rule_for_sd_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["pydantic"] = "pydantic"
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["pydantic"] = "pydantic"
     ) -> DisassociateAssociatedBudgetRuleResponse: ...
     @overload
     async def disassociate_associated_budget_rule_for_sd_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["dict"]
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["dict"]
     ) -> dict[str, Any]: ...
     @overload
     async def disassociate_associated_budget_rule_for_sd_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["raw"]
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["raw"]
     ) -> httpx.Response: ...
     async def disassociate_associated_budget_rule_for_sd_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
     ) -> DisassociateAssociatedBudgetRuleResponse | dict[str, Any] | httpx.Response:
         """"""
 
@@ -173,18 +173,18 @@ class BudgetRules(BaseResource):
 
     @overload
     async def list_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, *, mode: Literal["pydantic"] = "pydantic"
+        self, campaign_id: int, *, mode: Literal["pydantic"] = "pydantic"
     ) -> SDListAssociatedBudgetRulesResponse: ...
     @overload
     async def list_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, *, mode: Literal["dict"]
+        self, campaign_id: int, *, mode: Literal["dict"]
     ) -> dict[str, Any]: ...
     @overload
     async def list_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, *, mode: Literal["raw"]
+        self, campaign_id: int, *, mode: Literal["raw"]
     ) -> httpx.Response: ...
     async def list_associated_budget_rules_for_sd_campaigns(
-        self, campaign_id: float, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
+        self, campaign_id: int, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
     ) -> SDListAssociatedBudgetRulesResponse | dict[str, Any] | httpx.Response:
         """**Requires one of these permissions**:"""
