+++ src/ads_api/client/v0/sp_v3/budget_rules.py
@@ -73,19 +73,19 @@ class BudgetRules(BaseResource):
 
     @overload
     async def create_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["pydantic"] = "pydantic"
+        self, campaign_id: int, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["pydantic"] = "pydantic"
     ) -> CreateAssociatedBudgetRulesResponse: ...
     @overload
     async def create_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["dict"]
+        self, campaign_id: int, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["dict"]
     ) -> dict[str, Any]: ...
     @overload
     async def create_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["raw"]
+        self, campaign_id: int, body: CreateAssociatedBudgetRulesRequest, *, mode: Literal["raw"]
     ) -> httpx.Response: ...
     async def create_associated_budget_rules_for_sp_campaigns(
         self,
-        campaign_id: float,
+        campaign_id: int,
         body: CreateAssociatedBudgetRulesRequest,
         *,
         mode: Literal["pydantic", "dict", "raw"] = "pydantic",
@@ -117,18 +117,18 @@ class BudgetRules(BaseResource):
 
     @overload
     async def disassociate_associated_budget_rule_for_sp_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["pydantic"] = "pydantic"
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["pydantic"] = "pydantic"
     ) -> DisassociateAssociatedBudgetRuleResponse: ...
     @overload
     async def disassociate_associated_budget_rule_for_sp_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["dict"]
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["dict"]
     ) -> dict[str, Any]: ...
     @overload
     async def disassociate_associated_budget_rule_for_sp_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["raw"]
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["raw"]
     ) -> httpx.Response: ...
     async def disassociate_associated_budget_rule_for_sp_campaigns(
-        self, campaign_id: float, budget_rule_id: str, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
+        self, campaign_id: int, budget_rule_id: str, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
     ) -> DisassociateAssociatedBudgetRuleResponse | dict[str, Any] | httpx.Response:
         """**Requires one of these permissions**:"""
 
@@ -217,18 +217,18 @@ class BudgetRules(BaseResource):
 
     @overload
     async def list_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, *, mode: Literal["pydantic"] = "pydantic"
+        self, campaign_id: int, *, mode: Literal["pydantic"] = "pydantic"
     ) -> SPListAssociatedBudgetRulesResponse: ...
     @overload
     async def list_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, *, mode: Literal["dict"]
+        self, campaign_id: int, *, mode: Literal["dict"]
     ) -> dict[str, Any]: ...
     @overload
     async def list_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, *, mode: Literal["raw"]
+        self, campaign_id: int, *, mode: Literal["raw"]
     ) -> httpx.Response: ...
     async def list_associated_budget_rules_for_sp_campaigns(
-        self, campaign_id: float, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
+        self, campaign_id: int, *, mode: Literal["pydantic", "dict", "raw"] = "pydantic"
     ) -> SPListAssociatedBudgetRulesResponse | dict[str, Any] | httpx.Response:
         """**Authorized resource type**:"""
